fix: use gradient from ManifoldDiff
