+++ test/natural_manifolds/bernoulli_tests.jl
@@ -7,11 +7,7 @@
 end
 
 @testitem "Check MLE works for `Bernoulli`" begin
-    include("natural_manifolds_setuptests.jl")
-
-    using Manopt
-    import Distributions: kldivergence, Distribution
-
+    include("mle_manifolds_setuptests.jl")
     test_mle_works(; mle_samples=1000, ndistributions=3) do rng
         return Bernoulli(rand(rng))
     end
