+++ test/natural_manifolds/beta_tests.jl
@@ -7,11 +7,7 @@
 end
 
 @testitem "Check MLE works for `Beta`" begin
-    include("natural_manifolds_setuptests.jl")
-    using Manopt
-    import Distributions: kldivergence, Distribution
-
-    # Use fewer samples/iterations for faster tests
+    include("mle_manifolds_setuptests.jl")
     test_mle_works(; mle_samples=1000, ndistributions=3) do rng
         return Beta(10rand(rng), 10rand(rng))
     end
