+++ test/natural_manifolds/categorical_tests.jl
@@ -58,10 +58,7 @@ end
 end
 
 @testitem "Check MLE works for `Categorical`" begin
-    include("natural_manifolds_setuptests.jl")
-
-    using Manopt
-    import Distributions: kldivergence, Distribution
+    include("mle_manifolds_setuptests.jl")
 
     @test_broken test_mle_works(; mle_samples=1000, ndistributions=3) do rng
         return Categorical(rand(rng, 3))
