+++ test/natural_manifolds/dirichlet_tests.jl
@@ -8,10 +8,7 @@
 end
 
 @testitem "Check MLE works for `Dirichlet`" begin
-    include("natural_manifolds_setuptests.jl")
-    using Manopt
-    import Distributions: kldivergence, Distribution
-
+    include("mle_manifolds_setuptests.jl")
     # Use fewer samples/iterations for faster tests
     test_mle_works(; mle_samples=500, ndistributions=3) do rng
         return Dirichlet(10rand(rng, 3))
