+++ test/natural_manifolds/exponential_tests.jl
@@ -7,10 +7,7 @@
 end
 
 @testitem "Check MLE works for `Exponential`" begin
-    include("natural_manifolds_setuptests.jl")
-    using Manopt
-    import Distributions: kldivergence, Distribution
-
+    include("mle_manifolds_setuptests.jl")
     test_mle_works(; mle_samples=500, ndistributions=3) do rng
         return Exponential(rand(rng))
     end
