+++ test/natural_manifolds/laplace_tests.jl
@@ -7,9 +7,9 @@
 end
 
 @testitem "Check MLE works for `Laplace`" begin
-    include("natural_manifolds_setuptests.jl")
+    include("mle_manifolds_setuptests.jl")
 
-    test_natural_manifold() do rng
+    test_mle_works(; mle_samples=500, ndistributions=3) do rng
         return Laplace(rand(rng), rand(rng))
     end
 end
