+++ test/natural_manifolds/negative_binomial_tests.jl
@@ -7,10 +7,7 @@
 end
 
 @testitem "Check MLE works for `NegativeBinomial`" begin
-    include("natural_manifolds_setuptests.jl")
-    using Manopt
-    import Distributions: kldivergence, Distribution
-
+    include("mle_manifolds_setuptests.jl")
     @test_broken test_mle_works(; mle_samples=1000, ndistributions=3) do rng
         return NegativeBinomial(rand(rng), rand(rng, 0.00001:0.01:1))
     end
