+++ test/natural_manifolds/weibull_tests.jl
@@ -7,10 +7,7 @@
 end
 
 @testitem "Check MLE works for `Weibull`" begin
-    include("natural_manifolds_setuptests.jl")
-    using Manopt
-    import Distributions: kldivergence, Distribution
-
+    include("mle_manifolds_setuptests.jl")
     test_mle_works(; mle_samples=500, ndistributions=3) do rng
         return Weibull(rand(rng), rand(rng))
     end
