+++ test/natural_manifolds/wishart_tests.jl
@@ -10,13 +10,11 @@
 end
 
 @testitem "Check MLE works for `Wishart`" begin
-    include("natural_manifolds_setuptests.jl")
+    include("mle_manifolds_setuptests.jl")
 
-    using Manopt
-    import Distributions: kldivergence, Distribution
     import ExponentialFamily: WishartFast
-
-    test_mle_works(; mle_samples=500, ndistributions=1, kl_friendly=false) do rng
+    
+    test_mle_works(; mle_samples=5000, ndistributions=1, kl_friendly=false) do rng
         k = 2
         L = LowerTriangular(randn(rng, k, k))
         C = L * L' + k * I
